feat: add responsive sidebar with active link highlighting
